+++ carp_mobile_sensing/test/core_json_test.dart
@@ -4,7 +4,8 @@ import 'dart:io';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:test/test.dart';
 
-String _encode(Object object) => const JsonEncoder.withIndent(' ').convert(object);
+String _encode(Object object) =>
+    const JsonEncoder.withIndent(' ').convert(object);
 
 void main() {
   Study study;
@@ -14,7 +15,7 @@ void main() {
     //SamplingPackageRegistry.register(CommunicationSamplingPackage());
     //SamplingPackageRegistry.register(ContextSamplingPackage());
 
-    study = Study("1234", "bardram", name: "bardram study");
+    study = Study("1234", "bardram", name: "bardram study", deploymentId: '#1');
     //study.dataEndPoint = DataEndPoint(DataEndPointType.PRINT);
     study.dataEndPoint = FileDataEndPoint()
       ..bufferSize = 50 * 1000
@@ -25,8 +26,10 @@ void main() {
     study.addTriggerTask(
         ImmediateTrigger(), // a simple trigger that starts immediately
         AutomaticTask(name: 'Sampling Task')
-          ..measures =
-              SamplingSchema.common(namespace: NameSpace.CARP).measures.values.toList() // a task with all measures
+          ..measures = SamplingSchema.common(namespace: NameSpace.CARP)
+              .measures
+              .values
+              .toList() // a task with all measures
         );
   });
 
@@ -34,13 +37,16 @@ void main() {
     print(_encode(study));
 
     expect(study.id, "1234");
+    expect(study.deploymentId, "#1");
   });
 
   test('JSON -> Study, assert study id', () async {
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(study_2.id, study.id);
+    expect(study.deploymentId, "#1");
 
     print(_encode(study_2));
   });
@@ -48,14 +54,16 @@ void main() {
   test('JSON -> Study, deep assert', () async {
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(_encode(study_2), equals(studyJson));
   });
 
   test('Configuration -> JSON', () async {
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(study_2.name, study.name);
   });
 
@@ -64,18 +72,20 @@ void main() {
     String plainStudyJson = File("test/study_1234.json").readAsStringSync();
     print(plainStudyJson);
 
-    Study plainStudy = Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
+    Study plainStudy =
+        Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
     expect(plainStudy.id, study.id);
 
     final studyJson = _encode(study);
 
-    Study study_2 = Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
+    Study study_2 =
+        Study.fromJson(json.decode(plainStudyJson) as Map<String, dynamic>);
     expect(_encode(study_2), equals(studyJson));
   });
 
   test('Data point -> JSON', () async {
-    var dp =
-        DataPoint.fromDatum(study.id, study.userId, MapDatum({'item_1': '12.23423452345', 'item_2': '3.82375823475'}));
+    var dp = DataPoint.fromDatum(study.id, study.userId,
+        MapDatum({'item_1': '12.23423452345', 'item_2': '3.82375823475'}));
     print(_encode(dp));
 
     BatteryDatum datum = BatteryDatum()
@@ -97,61 +107,80 @@ void main() {
     study_3.addTriggerTask(
         DelayedTrigger(delay: Duration(seconds: 10)),
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures = SamplingSchema.common()
-              .getMeasureList(types: [SensorSamplingPackage.PEDOMETER, DeviceSamplingPackage.SCREEN]));
+          ..measures = SamplingSchema.common().getMeasureList(types: [
+            SensorSamplingPackage.PEDOMETER,
+            DeviceSamplingPackage.SCREEN
+          ]));
 
     study_3.addTriggerTask(
-        PeriodicTrigger(period: const Duration(minutes: 1)), // collect every min.
+        PeriodicTrigger(
+            period: const Duration(minutes: 1)), // collect every min.
         AutomaticTask(name: 'Sensing Task #2')
-          ..measures = SamplingSchema.common()
-              .getMeasureList(types: [SensorSamplingPackage.LIGHT, DeviceSamplingPackage.DEVICE]));
+          ..measures = SamplingSchema.common().getMeasureList(types: [
+            SensorSamplingPackage.LIGHT,
+            DeviceSamplingPackage.DEVICE
+          ]));
 
     RecurrentScheduledTrigger t1, t2, t3, t4;
 
     // collect every day at 13:30.
-    t1 = RecurrentScheduledTrigger(type: RecurrentType.daily, time: Time(hour: 21, minute: 30));
+    t1 = RecurrentScheduledTrigger(
+        type: RecurrentType.daily, time: Time(hour: 21, minute: 30));
     print('$t1');
     study_3.addTriggerTask(
         t1,
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures = SamplingSchema.common().getMeasureList(types: [DeviceSamplingPackage.MEMORY]));
+          ..measures = SamplingSchema.common()
+              .getMeasureList(types: [DeviceSamplingPackage.MEMORY]));
 
     // collect every other day at 13:30.
-    t2 = RecurrentScheduledTrigger(type: RecurrentType.daily, time: Time(hour: 13, minute: 30), separationCount: 1);
+    t2 = RecurrentScheduledTrigger(
+        type: RecurrentType.daily,
+        time: Time(hour: 13, minute: 30),
+        separationCount: 1);
     print('$t2');
     study_3.addTriggerTask(
         t2,
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures =
-              SamplingSchema.common().getMeasureList(types: [AppsSamplingPackage.APPS, DeviceSamplingPackage.MEMORY]));
+          ..measures = SamplingSchema.common().getMeasureList(
+              types: [AppsSamplingPackage.APPS, DeviceSamplingPackage.MEMORY]));
 
     // collect every wednesday at 12:23.
     t3 = RecurrentScheduledTrigger(
-        type: RecurrentType.weekly, time: Time(hour: 12, minute: 23), dayOfWeek: DateTime.wednesday);
+        type: RecurrentType.weekly,
+        time: Time(hour: 12, minute: 23),
+        dayOfWeek: DateTime.wednesday);
     print('$t3');
     study_3.addTriggerTask(
         t3,
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures =
-              SamplingSchema.common().getMeasureList(types: [AppsSamplingPackage.APPS, AppsSamplingPackage.APP_USAGE]));
+          ..measures = SamplingSchema.common().getMeasureList(types: [
+            AppsSamplingPackage.APPS,
+            AppsSamplingPackage.APP_USAGE
+          ]));
 
     // collect every 2nd monday at 12:23.
     t4 = RecurrentScheduledTrigger(
-        type: RecurrentType.weekly, time: Time(hour: 12, minute: 23), dayOfWeek: DateTime.monday, separationCount: 1);
+        type: RecurrentType.weekly,
+        time: Time(hour: 12, minute: 23),
+        dayOfWeek: DateTime.monday,
+        separationCount: 1);
     print('$t4');
     study_3.addTriggerTask(
         t4,
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures =
-              SamplingSchema.common().getMeasureList(types: [AppsSamplingPackage.APPS, DeviceSamplingPackage.MEMORY]));
+          ..measures = SamplingSchema.common().getMeasureList(
+              types: [AppsSamplingPackage.APPS, DeviceSamplingPackage.MEMORY]));
 
     // when battery level is 10% then find the list of apps running
     study_3.addTriggerTask(
         SamplingEventTrigger(
-            measureType: MeasureType(NameSpace.CARP, DeviceSamplingPackage.BATTERY),
+            measureType:
+                MeasureType(NameSpace.CARP, DeviceSamplingPackage.BATTERY),
             resumeCondition: BatteryDatum()..batteryLevel = 10),
         AutomaticTask(name: 'Sensing Task #1')
-          ..measures = SamplingSchema.common().getMeasureList(types: [AppsSamplingPackage.APP_USAGE]));
+          ..measures = SamplingSchema.common()
+              .getMeasureList(types: [AppsSamplingPackage.APP_USAGE]));
 
 // don't add ConditionalSamplingEventTrigger since it cannot be serialized to JSON
 //    study_3.addTriggerTask(
@@ -165,7 +194,8 @@ void main() {
 
     print(studyJson);
 
-    Study study_4 = Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    Study study_4 =
+        Study.fromJson(json.decode(studyJson) as Map<String, dynamic>);
     expect(study_4.id, study_3.id);
   });
 }